Remove duplicate JsMaterialXView and test
JsMaterialXView is renamed as MaterialXView in javascript folder
test            is renamed as MaterialXTest in javascript folder
