--- source/JsMaterialX/test/.babelrc
@@ -1,10 +1,0 @@
-{
-    "presets": [
-        [
-          "@babel/preset-env",
-          {
-            "exclude": ["transform-regenerator"]
-          }
-        ]
-      ]
-}
